+++ apps/api/src/auth/domain/model/action-request/event/action-request-was-created.ts
@@ -1,6 +1,10 @@
 import { DomainEventInterface } from '~/common/domain/model/event/domain-event.interface';
 import { DtoInterface } from '~/common/application/query/dto/dto.interface';
-import { ActionDataSerialized, ActionType, ActionStatus } from '~/auth/domain/model/action-request/action-request.model';
+import {
+  ActionDataSerialized,
+  ActionType,
+  ActionStatus,
+} from '~/auth/domain/model/action-request/action-request.model';
 
 export class ActionRequestWasCreated implements DomainEventInterface {
   eventName = 'auth.actionRequest.wasCreated';