+++ apps/api/src/main.ts
@@ -1,13 +1,21 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { VersioningType } from '@nestjs/common';
+import { ConsoleLogger, VersioningType } from '@nestjs/common';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
+  const app = await NestFactory.create(AppModule, {
+    logger: new ConsoleLogger({
+      json: true,
+    }),
+  });
   app.enableCors();
   app.enableVersioning({
     type: VersioningType.URI,
     defaultVersion: '1',
+  });
+  app.enableCors({
+    origin: true,
+    credentials: true,
   });
 
   await app.listen(3001);